[NX-448] Component Storage (nexus-asset-store)
