--- components/nexus-component-model/src/main/java/org/sonatype/nexus/component/model/Component.java
@@ -1,40 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2007-2014 Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package org.sonatype.nexus.component.model;
-
-import java.util.Map;
-
-import org.sonatype.nexus.componentmetadata.Record;
-
-/**
- * Represents an identifiable component.
- * 
- * @since 3.0
- */
-public interface Component
-{
-  /**
-   * @return canonical id of this component
-   */
-  ComponentId getId();
-
-  /**
-   * @return current metadata record for this component
-   */
-  Record getMetadata();
-
-  /**
-   * @return assets belonging to this component, by asset path
-   */
-  Map<String, Asset> getAssets();
-}
