+++ components/nexus-component-recordstore/src/main/java/org/sonatype/nexus/component/recordstore/package-info.java
@@ -10,17 +10,9 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.component.model;
-
 /**
- * Opaque component id in the normalized component system.
- * 
+ * The component record store api.
+ *
  * @since 3.0
  */
-public interface ComponentId
-{
-  /**
-   * Gets the id as a unique string.
-   */
-  String asUniqueString();
-}
+package org.sonatype.nexus.component.recordstore;
