+++ plugins/basic/nexus-rawcontent-plugin/pom.xml
@@ -45,11 +45,6 @@
 
     <dependency>
       <groupId>org.sonatype.nexus</groupId>
-      <artifactId>nexus-componentmetadata</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.sonatype.nexus</groupId>
       <artifactId>nexus-blobstore-file</artifactId>
     </dependency>
