+++ pom.xml
@@ -190,7 +190,19 @@
 
       <dependency>
         <groupId>org.sonatype.nexus</groupId>
-        <artifactId>nexus-componentmetadata</artifactId>
+        <artifactId>nexus-component-assetstore</artifactId>
+        <version>3.0.0-SNAPSHOT</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.sonatype.nexus</groupId>
+        <artifactId>nexus-component-model</artifactId>
+        <version>3.0.0-SNAPSHOT</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.sonatype.nexus</groupId>
+        <artifactId>nexus-component-recordstore</artifactId>
         <version>3.0.0-SNAPSHOT</version>
       </dependency>
